+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 All major and minor version changes will be documented in this file. Details of
 patch-level version changes can be found in [commit messages](../../commits/master).
 
+## 2020.2.1 - 2020/04/23
+- Fixed incorrect docstrings
+- Added SOFTLIGHT and HARDLIGHT
+
 ## 2020.2 - 2020/04/23
 - Added basic support for blend modes NORMAL, MULTIPLY, ADDITIVE, COLOURBURN,
 COLOURDODGE, REFLECT, GLOW, OVERLAY, DIFFERENCE, NEGATION, LIGHTEN, DARKEN,